updated quiz

+++ src/main/java/DAO/StandardUnorderedResponseQuestionDao.java
@@ -34,12 +34,12 @@ public class StandardUnorderedResponseQuestionDao implements QuestionDao {
     }
 
     @Override
-    public List<Question> getQuestions(int quizId) {
+    public List<Question> getQuestions(long quizId) {
         List<Question> result = new ArrayList<>();
 
         try {
             PreparedStatement st = conn.prepareStatement("select * from standard_unordered_questions WHERE  quiz_id = ?;" );
-            st.setInt(1, quizId);
+            st.setLong(1, quizId);
             ResultSet res = st.executeQuery();
 
             while(res.next()){
@@ -55,20 +55,20 @@ public class StandardUnorderedResponseQuestionDao implements QuestionDao {
         }
         return result;
     }
-    private void insertAnswers(String st, Connection conn, int question_id, HashSet<String> answers) throws SQLException {
+    private void insertAnswers(String st, Connection conn, long question_id, HashSet<String> answers) throws SQLException {
         for(String s : answers){
             PreparedStatement statement1 = conn.prepareStatement(st);
             statement1.setString(1, s);
-            statement1.setInt(2, question_id);
+            statement1.setLong(2, question_id);
             statement1.execute();
         }
     }
 
 
-    private HashSet<String> getAnswers(int question_id, String s, Connection conn) throws SQLException {
+    private HashSet<String> getAnswers(long question_id, String s, Connection conn) throws SQLException {
         HashSet<String> result = new HashSet<>();
         PreparedStatement st = conn.prepareStatement(s);
-        st.setInt(1, question_id);
+        st.setLong(1, question_id);
         ResultSet res = st.executeQuery();
         while(res.next()){
             result.add(res.getString("answer_text"));
